pyproject.toml
[project]
name = "aia-t5-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "numpy>=2.4.4",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
]
